// frontend/src/app/features/caisse/models/categorie-operation-caisse.ts
export type CategorieOperationCaisse =
  | 'COTISATION'
  | 'EPARGNE'
  | 'RETRAIT_EPARGNE'
  | 'FRAIS_DEMANDE'
  | 'DEPOT_GARANTIE'
  | 'DECAISSEMENT_CREDIT'
  | 'REMBOURSEMENT_CREDIT'
  | 'PENALITE_RETARD'
  | 'APPROVISIONNEMENT'
  | 'DEPENSE'
  | 'ENTREE_DIVERSE'
  | 'SORTIE_DIVERSE'
  | 'FRAIS_DEMANDE_CREDIT'
  | 'FRAIS_RETRAIT_EPARGNE'
  | 'DEPOT_GARANTIE_CREDIT';